Always keep aspect scaling (lowest scalar of width and height) for fonts, lines, dots, spheres, reticle and non-Cockpit-/Statusbar-related HUD-elements; Moved the Cockpit alpha decoding to gauges.c and created a seperate bitmap for the alpha fields and create subbitmaps from this one - more memory but this way we do not modify the cockpit bitmap itself

+++ main/gamefont.h
@@ -46,9 +46,9 @@ COPYRIGHT 1993-1999 PARALLAX SOFTWARE CORPORATION.  ALL RIGHTS RESERVED.
 extern float FNTScaleX, FNTScaleY;
 
 // add (scaled) spacing to given font coordinate
-#define FSPACX(x)	((float)(x)*(FNTScaleX*(GAME_FONT->ft_w/7)))
-#define FSPACY(y)	((float)(y)*(FNTScaleY*(GAME_FONT->ft_h/5)))
-#define LINE_SPACING    ((float)FNTScaleY*(grd_curcanv->cv_font->ft_h+(GAME_FONT->ft_h/5)))
+#define FSPACX(x)	((int)((x)*(FNTScaleX*(GAME_FONT->ft_w/7))))
+#define FSPACY(y)	((int)((y)*(FNTScaleY*(GAME_FONT->ft_h/5))))
+#define LINE_SPACING    ((int)(FNTScaleY*(grd_curcanv->cv_font->ft_h+(GAME_FONT->ft_h/5))))
 
 extern grs_font *Gamefonts[MAX_FONTS];
